+++ laserbeamsize/display.py
@@ -5,45 +5,48 @@
 # pylint: disable=too-many-statements
 
 """
-A module for graphing and displaying the beam size fit.
+A module for generating a graphical analysis of beam size fitting.
 
-A full graphic can be created by::
+Full documentation is available at <https://laserbeamsize.readthedocs.io>
 
-    >>>> import imageio
-    >>>> import numpy as np
-    >>>> import laserbeamsize as lbs
-    >>>> beam_image = imageio.v2.imread("t-hene.pgm")
-    >>>> x, y, dx, dy, phi = lbs.beam_size(beam_image)
-    >>>> print("The center of the beam ellipse is at (%.0f, %.0f)" % (x, y))
-    >>>> print("The ellipse diameter (closest to horizontal) is %.0f pixels" % dx)
-    >>>> print("The ellipse diameter (closest to   vertical) is %.0f pixels" % dy)
-    >>>> print("The ellipse is rotated %.0f° ccw from the horizontal" % (phi * 180/3.1416))
-    >>>> lbs.beam_size_plot(beam_image)
-    >>>> plt.show()
+A graphic showing the image and extracted beam parameters is achieved by::
+
+    >>> import imageio.v3 as iio
+    >>> import matplotlib.pyplot as plt
+    >>> import laserbeamsize as lbs
+    >>>
+    >>> repo = "https://github.com/scottprahl/laserbeamsize/raw/master/docs/"
+    >>> image = iio.imread(repo + 't-hene.pgm')
+    >>>
+    >>> lbs.plot_image_analysis(image)
+    >>> plt.show()
 
 A mosaic of images might be created by::
 
-    >>>> # read images for each location
-    >>>> z = np.array([89,94,99,104,109,114,119,124,129,134,139], dtype=float) #[mm]
-    >>>> filenames = ["%d.pgm" % location for location in z]
-    >>>> images = [imageio.v2.imread(filename) for filename in filenames]
-    >>>> lbs.beam_size_montage(images, z * 1e-3, pixel_size=3.75, crop=True)
-    >>>> plt.show()
-
-Full documentation is available at <https://laserbeamsize.readthedocs.io>
+    >>> import imageio.v3 as iio
+    >>> import matplotlib.pyplot as plt
+    >>> import numpy as np
+    >>> import laserbeamsize as lbs
+    >>>
+    >>> repo = "https://github.com/scottprahl/laserbeamsize/raw/master/docs/"
+    >>> z1 = np.array([168,210,280,348,414,480], dtype=float)
+    >>> fn1 = [repo + "t-%dmm.pgm" % number for number in z1]
+    >>> images = [iio.imread(fn) for fn in fn1]
+    >>>
+    >>> options = {'z':z1/1000, 'pixel_size':0.00375, 'units':'mm', 'crop':True}
+    >>> lbs.plot_image_montage(images, **options, iso_noise=False)
+    >>> plt.show()
 """
 
 import numpy as np
 import matplotlib.pyplot as plt
-import laserbeamsize.image_tools as tools
-import laserbeamsize.background as back
-from laserbeamsize.analysis import beam_size
+import laserbeamsize as lbs
 
 __all__ = ('beam_ellipticity',
-           'draw_beam_figure',
-           'beam_size_plot',
-           'beam_size_and_plot',
-           'beam_size_montage'
+           'plot_beam_diagram',
+           'plot_image_analysis',
+           'plot_image_and_fit',
+           'plot_image_montage'
            )
 
 
@@ -78,13 +81,10 @@ def beam_ellipticity(dx, dy):
     return ellipticity, d_circular
 
 
-def draw_beam_figure():
+def plot_beam_diagram():
     """Draw a simple astigmatic beam ellipse with labels."""
     theta = np.radians(30)
-    xc = 0
-    yc = 0
-    dx = 50
-    dy = 25
+    xc, yc, dx, dy = 0, 0, 50, 25
 
     plt.subplots(1, 1, figsize=(6, 6))
 
@@ -92,10 +92,10 @@ def draw_beam_figure():
     # do not appear to be orthogonal to each other!
     plt.axes().set_aspect('equal')
 
-    xp, yp = tools.ellipse_arrays(xc, yc, dx, dy, theta)
+    xp, yp = lbs.ellipse_arrays(xc, yc, dx, dy, theta)
     plt.plot(xp, yp, 'k', lw=2)
 
-    xp, yp = tools.rotated_rect_arrays(xc, yc, dx, dy, theta)
+    xp, yp = lbs.rotated_rect_arrays(xc, yc, dx, dy, theta)
     plt.plot(xp, yp, ':b', lw=2)
 
     sint = np.sin(theta) / 2
@@ -122,13 +122,13 @@ def draw_beam_figure():
     plt.axis('off')
 
 
-def draw_visible_dotted_line(xpts, ypts):
+def plot_visible_dotted_line(xpts, ypts):
     """Draw a dotted line that is is visible against images."""
     plt.plot(xpts, ypts, '-', color='#FFD700')
     plt.plot(xpts, ypts, ':', color='#0057B8')
 
 
-def beam_size_and_plot(o_image,
+def plot_image_and_fit(o_image,
                        pixel_size=None,
                        vmin=None,
                        vmax=None,
@@ -138,7 +138,7 @@ def beam_size_and_plot(o_image,
                        cmap='gist_ncar',
                        corner_fraction=0.035,
                        nT=3,
-                       iso_noise=False,
+                       iso_noise=True,
                        **kwargs):
     """
     Plot the image, fitted ellipse, integration area, and semi-major/minor axes.
@@ -146,16 +146,16 @@ def beam_size_and_plot(o_image,
     If pixel_size is defined, then the returned measurements are in units of
     pixel_size.
 
-    This is helpful for creating a mosaics of all the images created for an
+    This function helpful when creating a mosaics of all images captured for an
     experiment.
+
+    If `crop==True` then the displayed image is cropped to the ISO 11146 integration
+    rectangle.
 
     If `crop` is a two parameter list `[v, h]` then `v` and `h` are
     interpreted as the vertical and horizontal sizes of the rectangle.  The
     size is in pixels unless `pixel_size` is specified.  In that case the
     rectangle sizes are in whatever units `pixel_size` is .
-
-    If `crop==True` then the displayed image is cropped to the ISO 11146 integration
-    rectangle.
 
     All cropping is done after analysis and therefosre only affects
     what is displayed.  If the image needs to be cropped before analysis
@@ -186,7 +186,7 @@ def beam_size_and_plot(o_image,
     bs_args['corner_fraction'] = corner_fraction
 
     # find center and diameters
-    xc, yc, dx, dy, phi = beam_size(o_image, **bs_args)
+    xc, yc, dx, dy, phi = lbs.beam_size(o_image, **bs_args)
 
     # establish scale and correct label
     if pixel_size is None:
@@ -202,9 +202,9 @@ def beam_size_and_plot(o_image,
         ymax = yc + crop[0] / 2 / scale
         xmin = xc - crop[1] / 2 / scale
         xmax = xc + crop[1] / 2 / scale
-        image, xc, yc = tools.crop_image_to_rect(o_image, xc, yc, xmin, xmax, ymin, ymax)
+        image, xc, yc = lbs.crop_image_to_rect(o_image, xc, yc, xmin, xmax, ymin, ymax)
     elif crop:
-        image, xc, yc = tools.crop_image_to_integration_rect(o_image, xc, yc, dx, dy, phi)
+        image, xc, yc = lbs.crop_image_to_integration_rect(o_image, xc, yc, dx, dy, phi)
     else:
         image = o_image
 
@@ -224,16 +224,16 @@ def beam_size_and_plot(o_image,
     plt.ylabel(label)
 
     # draw semi-major and semi-minor axes
-    xp, yp = tools.axes_arrays(xc, yc, dx, dy, phi)
-    draw_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
+    xp, yp = lbs.axes_arrays(xc, yc, dx, dy, phi)
+    plot_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
 
     # show ellipse around beam
-    xp, yp = tools.ellipse_arrays(xc, yc, dx, dy, phi)
-    draw_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
+    xp, yp = lbs.ellipse_arrays(xc, yc, dx, dy, phi)
+    plot_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
 
     # show integration area around beam
-    xp, yp = tools.rotated_rect_arrays(xc, yc, dx, dy, phi)
-    draw_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
+    xp, yp = lbs.rotated_rect_arrays(xc, yc, dx, dy, phi)
+    plot_visible_dotted_line((xp - xc) * scale, (yp - yc) * scale)
 
     # set limits on axes
     plt.xlim(-xc * scale, (h - xc) * scale)
@@ -247,16 +247,16 @@ def beam_size_and_plot(o_image,
     return xc * scale, yc * scale, dx * scale, dy * scale, phi
 
 
-def beam_size_plot(o_image,
-                   title='Original',
-                   pixel_size=None,
-                   units='µm',
-                   crop=False,
-                   cmap='gist_ncar',
-                   corner_fraction=0.035,
-                   nT=3,
-                   iso_noise=False,
-                   **kwargs):
+def plot_image_analysis(o_image,
+                        title='Original',
+                        pixel_size=None,
+                        units='µm',
+                        crop=False,
+                        cmap='gist_ncar',
+                        corner_fraction=0.035,
+                        nT=3,
+                        iso_noise=True,
+                        **kwargs):
     """
     Create a visual report for image fitting.
 
@@ -274,7 +274,7 @@ def beam_size_plot(o_image,
 
     Args:
         o_image: 2D image of laser beam
-        title: optional title for upper left plot
+        title: (optional) title for upper left plot
         pixel_size: (optional) size of pixels
         units: (optional) string used for units used on axes
         crop: (optional) crop image to integration rectangle
@@ -283,14 +283,13 @@ def beam_size_plot(o_image,
         nothing
     """
     # only pass along arguments that apply to beam_size()
-    beamsize_keys = ['mask_diameters', 'max_iter', 'phi', 'iso_noise']
-    bs_args = dict((k, kwargs[k]) for k in beamsize_keys if k in kwargs)
+    bs_args = dict((k, kwargs[k]) for k in ['mask_diameters', 'max_iter', 'phi'] if k in kwargs)
     bs_args['iso_noise'] = iso_noise
     bs_args['nT'] = nT
     bs_args['corner_fraction'] = corner_fraction
 
     # find center and diameters
-    xc, yc, dx, dy, phi = beam_size(o_image, **bs_args)
+    xc, yc, dx, dy, phi = lbs.beam_size(o_image, **bs_args)
 
     # determine scaling and labels
     if pixel_size is None:
@@ -309,16 +308,16 @@ def beam_size_plot(o_image,
         ymax = yc + crop[0] / 2 / scale
         xmin = xc - crop[1] / 2 / scale
         xmax = xc + crop[1] / 2 / scale
-        image, xc, yc = tools.crop_image_to_rect(o_image, xc, yc, xmin, xmax, ymin, ymax)
+        image, xc, yc = lbs.crop_image_to_rect(o_image, xc, yc, xmin, xmax, ymin, ymax)
     elif crop:
-        image, xc, yc = tools.crop_image_to_integration_rect(o_image, xc, yc, dx, dy, phi)
+        image, xc, yc = lbs.crop_image_to_integration_rect(o_image, xc, yc, dx, dy, phi)
     else:
         image = o_image
 
     # subtract background
-    working_image = back.subtract_image_background(image, corner_fraction=corner_fraction,
-                                                   nT=nT, iso_noise=iso_noise)
-    bkgnd, _ = back.iso_background(image, corner_fraction=corner_fraction, nT=nT)
+    working_image = lbs.subtract_iso_background(image, corner_fraction=corner_fraction,
+                                                nT=nT, iso_noise=iso_noise)
+    bkgnd, _ = lbs.iso_background(image, corner_fraction=corner_fraction, nT=nT)
 
     min_ = image.min()
     max_ = image.max()
@@ -354,14 +353,14 @@ def beam_size_plot(o_image,
     plt.subplot(2, 2, 2)
     extent = np.array([-xc_s, h_s - xc_s, v_s - yc_s, -yc_s])
     im = plt.imshow(working_image, extent=extent, cmap=cmap)
-    xp, yp = tools.ellipse_arrays(xc, yc, dx, dy, phi) * scale
-    draw_visible_dotted_line(xp - xc_s, yp - yc_s)
+    xp, yp = lbs.ellipse_arrays(xc, yc, dx, dy, phi) * scale
+    plot_visible_dotted_line(xp - xc_s, yp - yc_s)
 
-    xp, yp = tools.axes_arrays(xc, yc, dx, dy, phi) * scale
-    draw_visible_dotted_line(xp - xc_s, yp - yc_s)
+    xp, yp = lbs.axes_arrays(xc, yc, dx, dy, phi) * scale
+    plot_visible_dotted_line(xp - xc_s, yp - yc_s)
 
-    xp, yp = tools.rotated_rect_arrays(xc, yc, dx, dy, phi) * scale
-    draw_visible_dotted_line(xp - xc_s, yp - yc_s)
+    xp, yp = lbs.rotated_rect_arrays(xc, yc, dx, dy, phi) * scale
+    plot_visible_dotted_line(xp - xc_s, yp - yc_s)
 
     plt.colorbar(im, fraction=0.046 * v_s / h_s, pad=0.04)
 #    plt.clim(min_, max_)
@@ -372,7 +371,7 @@ def beam_size_plot(o_image,
     plt.title('Image w/o background, center at (%.0f, %.0f) %s' % (xc_s, yc_s, units))
 
     # plot of values along semi-major axis
-    _, _, z, s = tools.major_axis_arrays(o_image, xc, yc, dx, dy, phi)
+    _, _, z, s = lbs.major_axis_arrays(image, xc, yc, dx, dy, phi)
     a = np.sqrt(2 / np.pi) / r_major * abs(np.sum(z - bkgnd) * (s[1] - s[0]))
     baseline = a * np.exp(-2) + bkgnd
 
@@ -384,14 +383,14 @@ def beam_size_plot(o_image,
     plt.annotate('', (-r_mag_s, baseline), (r_mag_s, baseline),
                  arrowprops={'arrowstyle': '<->'})
     plt.text(0, 1.1 * baseline, 'dx=%.0f %s' % (d_mag_s, units), va='bottom', ha='center')
-    plt.text(0, a, '  Gaussian Fit')
+    plt.text(0, bkgnd + a, '  Gaussian Fit')
     plt.xlabel('Distance from Center [%s]' % units)
     plt.ylabel('Pixel Intensity Along Semi-Major Axis')
     plt.title('Semi-Major Axis')
     # plt.gca().set_ylim(bottom=0)
 
     # plot of values along semi-minor axis
-    _, _, z, s = tools.minor_axis_arrays(o_image, xc, yc, dx, dy, phi)
+    _, _, z, s = lbs.minor_axis_arrays(image, xc, yc, dx, dy, phi)
     a = np.sqrt(2 / np.pi) / r_minor * abs(np.sum(z - bkgnd) * (s[1] - s[0]))
     baseline = a * np.exp(-2) + bkgnd
 
@@ -403,7 +402,7 @@ def beam_size_plot(o_image,
     plt.annotate('', (-r_min_s, baseline), (r_min_s, baseline),
                  arrowprops={'arrowstyle': '<->'})
     plt.text(0, 1.1 * baseline, 'dy=%.0f %s' % (d_min_s, units), va='bottom', ha='center')
-    plt.text(0, a, '  Gaussian Fit')
+    plt.text(0, bkgnd + a, '  Gaussian Fit')
     plt.xlabel('Distance from Center [%s]' % units)
     plt.ylabel('Pixel Intensity Along Semi-Minor Axis')
     plt.title('Semi-Minor Axis')
@@ -413,19 +412,19 @@ def beam_size_plot(o_image,
     plt.subplots_adjust(wspace=0.3)
 
 
-def beam_size_montage(images,
-                      z=None,
-                      cols=3,
-                      pixel_size=None,
-                      vmax=None,
-                      vmin=None,
-                      units='µm',
-                      crop=False,
-                      cmap='gist_ncar',
-                      corner_fraction=0.035,
-                      nT=3,
-                      iso_noise=False,
-                      **kwargs):
+def plot_image_montage(images,
+                       z=None,
+                       cols=3,
+                       pixel_size=None,
+                       vmax=None,
+                       vmin=None,
+                       units='µm',
+                       crop=False,
+                       cmap='gist_ncar',
+                       corner_fraction=0.035,
+                       nT=3,
+                       iso_noise=True,
+                       **kwargs):
     """
     Create a beam size montage for a set of images.
 
@@ -487,7 +486,7 @@ def beam_size_montage(images,
         cb = not (vmax is None) and (i + 1 == cols)
 
         # plot the image and gather the beam diameters
-        _, _, dx[i], dy[i], _ = beam_size_and_plot(im, **options, colorbar=cb)
+        _, _, dx[i], dy[i], _ = plot_image_and_fit(im, **options, colorbar=cb)
 
         # add a title
         if units == 'mm':